합동 분산-공분산행렬. Pooled variance.

분산-공분산행렬이 범주에 관계없이 동일하다고 가정하고, 합동 분산-공분산행렬을 범주별로 주어진 관측치를 이용하여 추정한다. pooled_variance returns common variance.

pooled_variance(.data, .group_var, .xvar)

Arguments

.data
관측 데이터 프레임. A raw data frame.

.group_var
범주변수. A variable to group by.

.xvar
분산-공분산 행렬에 포함될 변수. One or more variables separated by commas within c(). Each variable needs to be a numeric column of .data.

Value

하나의 분산-공분산 행렬. A covariance matrix.

Examples

data(binaryclass2, package = "dmtr")
pooled_variance(binaryclass2, class, c(x1, x2))

#>          x1       x2
#> x1 3.885714 2.114286
#> x2 2.114286 2.457143